Separation game2048 class and drawBlock class
